requirements.txt
gradio=3.36.1
torch
transformers
pillow